Update README.md
Tried to follow assignment to best of my ability. On that note, let it be known I am very grateful the GitHub part of this assignment is only a third of the points and the other 2 portions are not. I believe this is explanation enough on the results of this portion. 
GitHub is not my friend.